services/workload-generator/requirements.txt
kafka-python==2.0.2
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.3
PyYAML==6.0.1
prometheus-client==0.19.0
